requirements.txt
# Python 3.11+ Required
# Lightweight, Compatible Versions
# Removed: TensorFlow (causes DLL issues on Windows)
streamlit==1.28.1
newsapi==0.1.1
requests==2.31.0
pandas==2.1.1
numpy==1.26.4
scikit-learn==1.3.1
nltk==3.8.1
textblob==0.17.1
transformers==4.36.0
torch==2.0.1
plotly==5.17.0
matplotlib==3.8.0
seaborn==0.13.0
wordcloud==1.9.3
joblib==1.3.2
python-dotenv==1.0.1
beautifulsoup4==4.12.2
Pillow==10.4.0
